# public/openapi.yaml
openapi: 3.0.3
info:
  title: DVD Rental API
  description: DVD Rental dataset API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /auth/refresh:
    get:
      tags:
        - Authorization
      security:
        - refresh_token: [ ]
      summary: Refresh token
      description: Sets new access_token and logged_in(HTTPOnly:false). refresh_token stays the same until expires
      operationId: RefreshToken
      responses:
        '201':
          $ref: '#/components/responses/AccessToken'
        '400':
          $ref: '#/components/responses/GenericError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/GenericError'
...
  /address/{id}:
    get:
      tags:
        - Address
      description: Get address by id
      operationId: GetOneAddress
      responses:
        '200':
          $ref: '#/components/responses/OneAddress'
        '400':
          $ref: '#/components/responses/GenericError'
        '422':
          $ref: '#/components/responses/GenericError'
    put:
      tags:
        - Address
      security:
        - access_token: [ ]
        - Authorization: [ ]
      description: Update address
      operationId: UpdateAddress
      requestBody:
        $ref: '#/components/requestBodies/AddressInput'
      responses:
        '201':
          $ref: '#/components/responses/OneAddress'
        '400':
          $ref: '#/components/responses/GenericError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/GenericError'
    delete:
      tags:
        - Address
      security:
        - access_token: [ ]
        - Authorization: [ ]
      description: Delete an address
      operationId: DeleteAddress
      responses:
        '200':
          description: Successfully deleted
          content:
            application/json:
              schema:
                type: integer
        '400':
          $ref: '#/components/responses/GenericError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/GenericError'
...
  /staff/{id}:
    get:
      tags:
        - Staff
      description: Get staff by id
      operationId: GetOneStaff
      responses:
        '200':
          $ref: '#/components/responses/OneStaff'
        '400':
          $ref: '#/components/responses/GenericError'
        '422':
          $ref: '#/components/responses/GenericError'
    put:
      tags:
        - Staff
      security:
        - access_token: [ ]
        - Authorization: [ ]
      description: Update staff
      operationId: UpdateStaff
      requestBody:
        $ref: '#/components/requestBodies/StaffInput'
      responses:
        '201':
          $ref: '#/components/responses/OneStaff'
        '400':
          $ref: '#/components/responses/GenericError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/GenericError'
    delete:
      tags:
        - Staff
      security:
        - access_token: [ ]
        - Authorization: [ ]
      description: Delete staff
      operationId: DeleteStaff
      responses:
        '201':
          $ref: '#/components/responses/OneStaff'
        '400':
          $ref: '#/components/responses/GenericError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/GenericError'
...
  /language/{id}:
    get:
      tags:
        - Language
      description: Get language by id
      operationId: GetOneLanguage
      responses:
        '200':
          $ref: '#/components/responses/OneLanguage'
        '400':
          $ref: '#/components/responses/GenericError'
        '422':
          $ref: '#/components/responses/GenericError'
    put:
      tags:
        - Language
      security:
        - access_token: [ ]
        - Authorization: [ ]
      description: Update language
      operationId: UpdateLanguage
      requestBody:
        $ref: '#/components/requestBodies/LanguageInput'
      responses:
        '201':
          $ref: '#/components/responses/OneLanguage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/GenericError'
        '422':
          $ref: '#/components/responses/GenericError'
    delete:
      tags:
        - Language
      security:
        - access_token: [ ]
        - Authorization: [ ]
      description: Delete language
      operationId: DeleteLanguage
      responses:
        '200':
          $ref: '#/components/responses/OneLanguage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/GenericError'
        '422':
          $ref: '#/components/responses/GenericError'
components:
  securitySchemes:
    access_token:
      description: |
        You have to have access token in access token cookie to authorize request
      type: apiKey
      name: access_token
      in: cookie
    refresh_token:
      description: |
        Refresh token is used to gain new access token
      type: apiKey
      name: refresh_token
      in: cookie
    Authorization:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Prefix the value with "Authorization" to indicate the custom authorization type
  responses:
    OneStaff:
      description: Staff
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Staff'
...
    OneLanguage:
      description: Language
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Language'
...
    OneAddress:
      description: One address
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Address'
...
    AccessToken:
      description: Access token
      headers:
        Set-Cookie:
          schema:
            type: string
            example: access_token httponly, logged_in boolean
      content:
        application/json:
          schema:
            type: object
            properties:
              access_token:
                type: string
    Unauthorized:
      description: Unauthorized
      content: { }
    GenericError:
      description: Unexpected error
      content: { }
...
  requestBodies:
    StaffInput:
      required: true
      description: Update and create staff
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StaffInput'
...
    LanguageInput:
      required: true
      description: Update and create language
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LanguageInput'
...
    AddressInput:
      required: true
      description: Create and update address
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AddressInput'
...
  schemas:
    Staff:
      required:
        - id
        - first_name
        - last_name
        - email
        - store_id
        - active
        - username
        - password
        - picture
        - address
        - updated_at
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          nullable: true
          example: example@mail.com
        store_id:
          type: integer
          nullable: true
          example: 1
        active:
          type: boolean
          example: false
        username:
          type: string
          example: bigDaddy
        password:
          type: string
          nullable: true
          example: 12345678
        picture:
          type: string
          format: binary
          nullable: true
          example: 'null'
        address:
          anyOf:
            - type: 'null'
            - $ref: '#/components/schemas/Address'
        updated_at:
          type: string
          format: date-time
          example: 01.02.2020
    StaffInput:
      required:
        - first_name
        - last_name
        - active
        - username
      type: object
      properties:
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          nullable: true
          example: johnydoe@mail.com
        store_id:
          type: number
          nullable: true
          example: 1
        active:
          type: boolean
          example: false
        username:
          type: string
          example: exceptional
        password:
          type: string
          nullable: true
          example: 12345678
        picture:
          type: string
          format: binary
          nullable: true
          example: null
        address_id:
          type: integer
          nullable: true
          example: 1
...
    Language:
      required:
        - id
        - name
        - updated_at
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: German
        updated_at:
          type: string
          format: date-time
          example: 02.02.2003
    LanguageInput:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: English
...
    Country:
      required:
        - id
        - country
        - updated_at
      type: object
      properties:
        id:
          type: integer
          example: 1
        country:
          type: string
          example: Russia
        updated_at:
          type: string
          format: date-time
          example: 12.03.2024
...
    City:
      required:
        - id
        - city
        - country
        - updated_at
      type: object
      properties:
        id:
          type: integer
          example: 1
        city:
          type: string
          example: Moscow
        country:
          anyOf:
            - type: 'null'
            - $ref: '#/components/schemas/Country'
          example: 'null'
        updated_at:
          type: string
          format: date-time
          example: 02.02.2024
...
    Address:
      required:
        - id
        - address
        - address2
        - district
        - city
        - country
        - postal_code
        - phone
        - updated_at
      type: object
      properties:
        id:
          type: integer
          example: 1
        address:
          type: string
          example: 1 Vancouver Road, Norwich",NR7 9UJ
        address2:
          type: string
          nullable: true
          example: 1 Dashwood Drive, Telford",TF1 3QW
        district:
          type: string
          example: California
        city:
          anyOf:
            - type: 'null'
            - $ref: '#/components/schemas/City'
          example: 'null'
        country:
          anyOf:
            - type: 'null'
            - $ref: '#/components/schemas/Country'
          example: 'null'
        postal_code:
          type: integer
          nullable: true
          example: 630526
        phone:
          type: string
          example: +79993337788
        updated_at:
          type: string
          format: date-time
          example: 02.02.2020
    AddressInput:
      required:
        - address
        - district
        - phone
      type: object
      properties:
        address:
          type: string
          example: 1 Vancouver Road, Norwich",NR7 9UJ
        address2:
          type: string
          nullable: true
          example: 1 Dashwood Drive, Telford",TF1 3QW
        district:
          type: string
          example: New York
        city_id:
          type: integer
          nullable: true
          example: 1
        postal_code:
          type: integer
          nullable: true
          example: 333888
        phone:
          type: string
          example: +73447882211
...